# VideoIn: add benchmarks and improve performance of crop_8bpp() and yuyv_to_grey()