This PR adds benchmarks using gbenchmark and improve performance of
crop_8bpp() and yuyv_to_grey().