@lucasdemarchi 2016-01-13
instead of going assembly, try enabling the compiler (via CXXFLAGS) to generate optimized instructions and compare to the ones without. For MinnowBoard Max that would be:
 `-march=corei7 -mtune=corei7 -mfpmath=sse -msse4.2`

For ARM I think this would be: `-mfpu=neon -ffast-math`