@lucasdemarchi 2016-01-13
heheh... just found my blog post back when I was at university: http://www.politreco.com/2009/07/11/sharing-some-homeworks/