@ricardotk 2016-01-14
> heheh... just found my blog post back when I was at university: http://www.politreco.com/2009/07/11/sharing-some-homeworks/

Nice blog post! Learned some interesting stuff there (:

But it's a pity that the YouTube videos aren't available anymore, I was very excited to see your skills in blender 3D :laughing: