@jberaud 2016-01-14
>For a 640x480 image to 64x64 crop its 16.9%
For a 640x480 image to 240x240 crop its 12.2%
For a 640x480 image to 640x480 crop its 2.6%

Oops I misread the dats. Then it is a big improvement.