@jberaud 2016-01-14
Regarding the crop calculation, I could add the reviewed flag.