@ricardotk 2016-01-14
>instead of going assembly, try enabling the compiler (via CXXFLAGS) to generate optimized instructions and compare to the ones without. For MinnowBoard Max that would be:
-march=corei7 -mtune=corei7 -mfpmath=sse -msse4.2

>For ARM I think this would be: -mfpu=neon -ffast-math

@lucasdemarchi we've got even better results with these flags (:

At least for the `yuyv_to_grey()` its very clear the gain in performance.

Without the flags:

Benchmark | Time (ns) | CPU(ns) | Iterations
---------------|--------------|-------------|-------------
BM_Crop8bpp/64/64      |   2655    |   2651    | 218750                                 
BM_Crop8bpp/240/240    |  33617   |   33667    |  20792                                 
BM_Crop8bpp/640/480    | 188022    | 188201   |    3684                                 
BM_YuyvToGrey/4k   |       1714   |    1717   |  403847                                 
BM_YuyvToGrey/75k     |   32148    |  32180   |   21649                                 
BM_YuyvToGrey/300k    |  128647   |  128753    |   5385 

With the flags:

Benchmark | Time (ns) | CPU(ns) | Iterations
---------------|--------------|-------------|-------------
BM_Crop8bpp/64/64     |          2644    |     2643  |   232103                                 
BM_Crop8bpp/240/240   |      33413  |     33346  |     20792                                 
BM_Crop8bpp/640/480    |   174583  |   174154     |    3962                                 
BM_YuyvToGrey/4k      |          1149     |    1150  |   599998                                 
BM_YuyvToGrey/75k      |      22135    |   22146   |    32813                                 
BM_YuyvToGrey/300k      |    85989   |    86140      |   7778